Added tool: ETCD manager
I've created a PR which has been merged into master:

https://github.com/etcd-io/etcd/pull/11044

But somehow this change didn't appear in the docs, even though a new version (3.4.0) has been released since. So I'm a bit confused now.  I'm opening this PR as suggested by @spzala (see the link above) and also because the "Edit this page" points in your docs points here.
If this is not right, please let me know what to do :)

Thanks in advance!