Add etcd version support policy
The details on the supported versions are missing in the doc.
etcd users has asked this question in the past. The details here
can provide an overview of etcd versioning policy.
Eventually we should make this verisoning page available more
easily but for now adding the details in the current page.